code now allows multiple pre-processing types
with this commit, we now allow multiple types of preprocessing, not just
the `dcraw -4 -d` and then manual demosaicing. Most of the code is
pretty agnostic as to what those looked like, except in the mtf
function, which iterates through them, and the preprocess_image
function, which will actually make the calls to dcraw if the images
aren't already there.

 - we restructure the data folder so the images all go in subfolders:
   raw, preprocessed, luminance, or mask_check. we add these subfolders
   to .gitignore

 - adds image_size to the pts_dict. since those are pixel numbers, they
   will depend on the size of the image they were grabbed from. we
   simply multiply them by the ratio between that value and the size of
   the image we're trying to extract the circles from. I wanted to
   assume these two values are integer multiples of each other, but I
   can't: different conversions from raw to not-raw algorithms result in
   slightly differently sized images (whether you use dcraw or Nikon's
   demosaicing)