overhauls how mtf.main works
mtf.main now returns a dataframe with all the info in it, which makes
it much easier to understand and easier to use. this also means that
mtf.mtf is now much simpler, since it doesn't need to do that silly
construction dance.

we rename some of the keys in metadata to be consistent with the
dataframe's