update to preprocess_all rule
we split the preprocess_all rule into three sub-rules, because the
intermediate files created by the vng and ahd demosaicing steps have
the same filename and I can't figure out how to change
that. therefore, we can't simultaneously create the vng and ahd
demosaic preprocessed file for the same raw file. if you run the jobs
one at a time, it doesn't matter, but this way you can run more of
them at a time, speeding it up